refactor to separate component and ditch timing based approach
