+++ app/src/components/v-form/form-field.vue
@@ -7,6 +7,7 @@ import { isEqual } from 'lodash';
 import { computed, inject, ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { fieldAnimationsKey } from './composables/use-field-animations';
+import FormFieldAnimations from './form-field-animations.vue';
 import FormFieldInterface from './form-field-interface.vue';
 import FormFieldLabel from './form-field-label.vue';
 import FormFieldMenu, { type MenuOptions } from './form-field-menu.vue';
@@ -50,15 +51,7 @@ const emit = defineEmits(['toggle-batch', 'toggle-raw', 'unset', 'update:modelVa
 const { t } = useI18n();
 
 const fieldAnimations = inject(fieldAnimationsKey);
-
 const animationKey = computed(() => fieldAnimations?.getAnimationKey(props.field.field));
-const animationDelay = computed(() => fieldAnimations?.getAnimationDelay(props.field.field) ?? 0);
-
-function onAnimationEnd(event: AnimationEvent) {
-	// Only handle the cursor-appear animation, ignore bubbled events from children
-	if (event.animationName !== 'cursor-appear') return;
-	fieldAnimations?.clearAnimation(props.field.field);
-}
 
 const isDisabled = computed(() => {
 	if (props.disabled) return true;
@@ -188,7 +181,6 @@ function useComputedValues() {
 				'field-updated': animationKey,
 			},
 		]"
-		:style="animationKey ? { '--animation-delay': `${animationDelay}ms` } : undefined"
 	>
 		<v-menu v-if="!isLabelHidden" :disabled="disabledMenu" placement="bottom-start" show-arrow arrow-placement="start">
 			<template #activator="{ toggle, active }">
@@ -266,9 +258,7 @@ function useComputedValues() {
 			<template v-else>{{ validationPrefix }}{{ validationMessage }}</template>
 		</small>
 
-		<div v-if="animationKey" :key="animationKey" class="field-cursor" @animationend="onAnimationEnd">
-			<v-icon name="magic_button" />
-		</div>
+		<form-field-animations v-if="animationKey" :field="field.field" />
 	</div>
 </template>
 
@@ -316,137 +306,6 @@ function useComputedValues() {
 	> * {
 		position: relative;
 		z-index: 1;
-	}
-
-	/* Border pulse */
-	&::before {
-		content: '';
-		position: absolute;
-		inset: -4px;
-		border-radius: var(--theme--border-radius);
-		pointer-events: none;
-		box-shadow: 0 0 0 2px var(--theme--primary);
-		opacity: 0;
-		animation: border-pulse 2s ease-out var(--animation-delay, 0ms);
-		z-index: 2;
-	}
-
-	/* Text shimmer overlay */
-	&::after {
-		content: '';
-		position: absolute;
-		inset: 0;
-		border-radius: var(--theme--border-radius);
-		pointer-events: none;
-		background: linear-gradient(
-			90deg,
-			transparent 0%,
-			transparent 25%,
-			var(--theme--primary-background) 50%,
-			transparent 75%,
-			transparent 100%
-		);
-		opacity: 0;
-		animation: text-shimmer-sweep 2s ease-out var(--animation-delay, 0ms);
-		background-size: 200% 100%;
-		mix-blend-mode: soft-light;
-		z-index: 3;
-		overflow: hidden;
-	}
-}
-
-@keyframes border-pulse {
-	0% {
-		opacity: 0;
-	}
-
-	10% {
-		opacity: 0.9;
-	}
-
-	25% {
-		opacity: 0.7;
-	}
-
-	75% {
-		opacity: 0.5;
-	}
-
-	90% {
-		opacity: 0.2;
-	}
-
-	100% {
-		opacity: 0;
-	}
-}
-
-@keyframes text-shimmer-sweep {
-	0% {
-		opacity: 0;
-		background-position: 250% 0;
-	}
-
-	10% {
-		opacity: 0.8;
-	}
-
-	20% {
-		opacity: 1;
-		background-position: 150% 0;
-	}
-
-	60% {
-		opacity: 1;
-		background-position: -50% 0;
-	}
-
-	70% {
-		opacity: 0.8;
-	}
-
-	100% {
-		opacity: 0;
-		background-position: -150% 0;
-	}
-}
-
-.field-cursor {
-	position: absolute;
-	inset-block-start: 0;
-	inset-inline-end: 0;
-	pointer-events: none;
-	z-index: 10;
-	color: var(--theme--primary);
-	font-size: 32px;
-	opacity: 0;
-	animation: cursor-appear 2s ease-out var(--animation-delay, 0ms);
-}
-
-@keyframes cursor-appear {
-	0% {
-		opacity: 0;
-		transform: scale(0.5) rotate(-10deg);
-	}
-
-	15% {
-		opacity: 1;
-		transform: scale(1.1) rotate(0deg);
-	}
-
-	25% {
-		opacity: 1;
-		transform: scale(1) rotate(0deg);
-	}
-
-	85% {
-		opacity: 1;
-		transform: scale(1) rotate(0deg);
-	}
-
-	100% {
-		opacity: 0;
-		transform: scale(0.8) rotate(5deg);
 	}
 }
 </style>
